rearranged functions
